requirements.txt
torch
torchvision
opencv-python
scipy
numpy
Pillow==8.4.0
scikit-image
bumpversion==0.5.3
wheel==0.29.0
watchdog==0.8.3
flake8
tox==2.3.1
coverage==4.1
Sphinx==1.4.8
cryptography==3.3.2
pyyaml>=4.2b1
face_recognition_models
Click>=6.0
dlib>=19.3.0
numpy
scipy